Define Board class, the basic unit of computation for custo
Board will serve as the basic unit of computation for future
Cutting Stock Problem algorithm implementations. Boards are cut into specific
sizes, optimized to reduce waste.

Board is initialized with material_size;
Board carries a contents, made of pieces which can be added or removed;
Board keeps a running total of the space left on it with space_remaining.

Future functionality would include BoardCollection, a grouping of Boards.