Implement and test BoardCollection class
Represents a collection of Boards, to serve as the result of our algorithm calculation. It will be needed in the future when we are implementing solutions to the Cutting Stock Problem.

BoardCollection initializes its contents to []
num_boards is a supported property
append method adds a Board to the collection

-> Tested via test_boardCollection.py