Refactor Board and BoardCollection tests to single file
This was done to have /tests structure match the module structure.

As an alternative, board.py could have been split into board.py and boardcollection.py. However, this path was not chosen because the classes Board and BoardCollection are logically connected. Therefore, they should remain in the same module. Thus, tests should also be completed in one file.